chg 日记模块点击不灵敏,评论模块样式调整

+++ app/src/main/java/me/jessyan/mvparms/demo/mvp/model/entity/user/bean/GrowthInfo.java
@@ -8,7 +8,7 @@ public class GrowthInfo {
 
     private String name;
     private long growth;
-    private long percent;
+    private double percent;
     private String createDate;
     private long inGrowth;
     private String type;
@@ -30,11 +30,11 @@ public class GrowthInfo {
         this.growth = growth;
     }
 
-    public long getPercent() {
+    public double getPercent() {
         return percent;
     }
 
-    public void setPercent(long percent) {
+    public void setPercent(double percent) {
         this.percent = percent;
     }
